chore: begins local api abstraction
